TiendaMascota/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
cx-Oracle==8.3.0
Django==4.0.5
djangorestframework==3.13.1
idna==3.3
marshmallow==2.21.0
Pillow==9.1.1
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
transbank-sdk==3.0.0
tzdata==2022.1
urllib3==1.26.9
